+++ R/pop_age.R
@@ -4,10 +4,10 @@
 #' `r lifecycle::badge("deprecated")`
 #'
 #' `pop_age()` is deprecated. To align a population table to a contact
-#' matrix's age groups, use [regroup_ages()].
+#' matrix's age groups, use [rebin_ages()].
 #'
 #' @return data frame of age-specific population data
-#' @inheritParams regroup_ages_numeric
+#' @inheritParams rebin_ages_numeric
 #' @param age.limits,pop.age.column,pop.column
 #'   `r lifecycle::badge("deprecated")` Use the underscore
 #'   versions (e.g., `age_limits`) instead.
@@ -24,7 +24,7 @@ pop_age <- function(
   pop.age.column = deprecated(),
   pop.column = deprecated()
 ) {
-  lifecycle::deprecate_warn("0.7.0", "pop_age()", "regroup_ages()")
+  lifecycle::deprecate_warn("0.7.0", "pop_age()", "rebin_ages()")
   chkDots(...)
 
   ## Handle deprecated arguments
@@ -50,7 +50,7 @@ pop_age <- function(
     "pop_age"
   )
 
-  regroup_ages_numeric(
+  rebin_ages_numeric(
     pop = pop,
     age_limits = age_limits,
     pop_age_column = pop_age_column,